+++ lib/fields/address.dart
@@ -7,6 +7,7 @@ import 'package:every_door/screens/editor/addr_chooser.dart';
 import 'package:flutter/material.dart';
 import 'package:every_door/models/field.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class AddressField extends PresetField {
   AddressField({required String label})
@@ -95,8 +96,9 @@ class _AddressInputState extends ConsumerState<AddressInput> {
 
   @override
   Widget build(BuildContext context) {
+    final loc = AppLocalizations.of(context)!;
     if (nearestAddresses.isEmpty) {
-      return Text('No addresses nearby');
+      return Text(loc.filterNoAddresses);
     }
 
     final current = StreetAddress.fromTags(widget.element.getFullTags());